requirements.txt
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
psycopg2-binary
pytest==7.4.2
pytest-django==4.5.2
Faker==19.6.2
factory-boy==3.3.0
python-dotenv
gunicorn
